requirements.txt
langchain==0.1.13
langchain-openai=0.1.1
openai==1.14.2
tiktoken==0.6.0
PyPDF2==3.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
streamlit==1.32.2
streamlit-extras
altair<5